refactor(workflow-tasks): move helm component into library
